requirements.txt
requests
bs4
selenium
python-dotenv
sentry-sdk==0.15.1
